refactor(lab): extract pure metrics into lab_metrics module
